requirements.txt
fastapi
uvicorn
openai==0.28.0
requests
gtts
geopy==2.4.1
geonamescache==2.0.0
python-dotenv
